+++ account-keeper-impl/src/main/java/com/dwarfeng/acckeeper/impl/bean/entity/HibernateCheckerInfo.java
@@ -1,0 +1,101 @@
+package com.dwarfeng.acckeeper.impl.bean.entity;
+
+import com.dwarfeng.acckeeper.sdk.util.Constraints;
+import com.dwarfeng.subgrade.sdk.bean.key.HibernateStringIdKey;
+import com.dwarfeng.subgrade.stack.bean.Bean;
+
+import javax.persistence.*;
+import java.util.Optional;
+
+@Entity
+@IdClass(HibernateStringIdKey.class)
+@Table(name = "tbl_checker_info")
+public class HibernateCheckerInfo implements Bean {
+
+    private static final long serialVersionUID = -2484369624757888597L;
+
+    // -----------------------------------------------------------主键-----------------------------------------------------------
+    @Id
+    @Column(name = "id", length = Constraints.LENGTH_ID, nullable = false, unique = true)
+    private String stringId;
+
+    // -----------------------------------------------------------主属性字段-----------------------------------------------------------
+    @Column(name = "type", length = Constraints.LENGTH_TYPE)
+    private String type;
+
+    @Column(name = "param", columnDefinition = "TEXT")
+    private String param;
+
+    @Column(name = "remark", length = Constraints.LENGTH_REMARK)
+    private String remark;
+
+    // -----------------------------------------------------------一对一-----------------------------------------------------------
+    @OneToOne(targetEntity = HibernateAccount.class)
+    @JoinColumns({ //
+            @JoinColumn(name = "id", referencedColumnName = "id", insertable = false, updatable = false), //
+    })
+    private HibernateAccount account;
+
+    public HibernateCheckerInfo() {
+    }
+
+    // -----------------------------------------------------------映射用属性区-----------------------------------------------------------
+    public HibernateStringIdKey getKey() {
+        return Optional.ofNullable(stringId).map(HibernateStringIdKey::new).orElse(null);
+    }
+
+    public void setKey(HibernateStringIdKey uuidKey) {
+        this.stringId = Optional.ofNullable(uuidKey).map(HibernateStringIdKey::getStringId).orElse(null);
+    }
+
+    // -----------------------------------------------------------常规属性区-----------------------------------------------------------
+    public String getStringId() {
+        return stringId;
+    }
+
+    public void setStringId(String stringId) {
+        this.stringId = stringId;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getParam() {
+        return param;
+    }
+
+    public void setParam(String param) {
+        this.param = param;
+    }
+
+    public String getRemark() {
+        return remark;
+    }
+
+    public void setRemark(String remark) {
+        this.remark = remark;
+    }
+
+    public HibernateAccount getAccount() {
+        return account;
+    }
+
+    public void setAccount(HibernateAccount account) {
+        this.account = account;
+    }
+
+    @Override
+    public String toString() {
+        return getClass().getSimpleName() + "(" +
+                "stringId = " + stringId + ", " +
+                "type = " + type + ", " +
+                "param = " + param + ", " +
+                "remark = " + remark + ", " +
+                "account = " + account + ")";
+    }
+}