+++ account-keeper-impl/src/main/java/com/dwarfeng/acckeeper/impl/bean/entity/HibernateMapper.java
@@ -44,6 +44,7 @@ public interface HibernateMapper {
 
     @Mapping(target = "stringId", ignore = true)
     @Mapping(target = "protectorInfo", ignore = true)
+    @Mapping(target = "checkerInfo", ignore = true)
     HibernateAccount accountToHibernate(Account account);
 
     @InheritInverseConfiguration
@@ -95,4 +96,17 @@ public interface HibernateMapper {
 
     @InheritInverseConfiguration
     ProtectDetailRecord protectDetailRecordFromHibernate(HibernateProtectDetailRecord hibernateProtectDetailRecord);
+
+    @Mapping(target = "stringId", ignore = true)
+    @Mapping(target = "account", ignore = true)
+    HibernateCheckerInfo checkerInfoToHibernate(CheckerInfo checkerInfo);
+
+    @InheritInverseConfiguration
+    CheckerInfo checkerInfoFromHibernate(HibernateCheckerInfo hibernateCheckerInfo);
+
+    @Mapping(target = "stringId", ignore = true)
+    HibernateCheckerSupport checkerSupportToHibernate(CheckerSupport checkerSupport);
+
+    @InheritInverseConfiguration
+    CheckerSupport checkerSupportFromHibernate(HibernateCheckerSupport hibernateCheckerSupport);
 }